Default this week's backward to last week's forward
A lot of people were confused about the placeholder.  It looked like
real data, and some pages showed it after the save, so they thouht it
had been saved to the database as their new backward FAB.  This lead to
concern when those notes didn't show up on the list.